Make retry feature explicit
Previously the retry feature was an implementation detail. It might
confuse users when configuring a timeout and the actual request will
take multitude of that in case of retries.

This commit adds a new constructor to configure an explicit retry
count. The old constructor's behaviour remains unchanged by using a
default of 3 retries. Use 0 (KM200.RETRY_DISABLED) to disable retries
completely.